requirements.txt
python-telegram-bot>=22.0
openai>=1.6.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
dateparser>=1.2.0
deutschland>=0.1.1 